Refactor handlers into their own files, and hierarchy mappings
