+++ handlers/filter-overview.go
@@ -1,0 +1,77 @@
+package handlers
+
+import (
+	"encoding/json"
+	"net/http"
+
+	"github.com/ONSdigital/dp-frontend-filter-dataset-controller/data"
+	"github.com/ONSdigital/dp-frontend-filter-dataset-controller/mapper"
+	"github.com/ONSdigital/go-ns/log"
+	"github.com/gorilla/mux"
+)
+
+// FilterOverview controls the render of the filter overview template
+// Contains stubbed data for now - page to be populated by the API
+func (f *Filter) FilterOverview(w http.ResponseWriter, req *http.Request) {
+	vars := mux.Vars(req)
+	filterID := vars["filterID"]
+
+	dims, err := f.fc.GetDimensions(filterID)
+	if err != nil {
+		log.ErrorR(req, err, nil)
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+
+	var dimensions []data.Dimension
+	for _, dim := range dims {
+		var vals data.DimensionValues
+		vals, err = f.fc.GetDimensionOptions(filterID, dim.Name)
+		if err != nil {
+			log.ErrorR(req, err, nil)
+			w.WriteHeader(http.StatusInternalServerError)
+			return
+		}
+		var values []string
+		for _, val := range vals.Items {
+			values = append(values, val.Name)
+		}
+
+		dimensions = append(dimensions, data.Dimension{
+			Name:   dim.Name,
+			Values: values,
+		})
+	}
+
+	filter, err := f.fc.GetJobState(filterID)
+	if err != nil {
+		log.ErrorR(req, err, nil)
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+
+	dataset, err := f.dc.GetDataset(filterID, filter.Edition, filter.Version)
+	if err != nil {
+		log.ErrorR(req, err, nil)
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+
+	p := mapper.CreateFilterOverview(dimensions, filter, dataset, filterID)
+
+	b, err := json.Marshal(p)
+	if err != nil {
+		log.Error(err, nil)
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+
+	templateBytes, err := f.r.Do("dataset-filter/filter-overview", b)
+	if err != nil {
+		log.Error(err, nil)
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+
+	w.Write(templateBytes)
+}
