+++ stubbed-apis/README.md
@@ -1,0 +1,1 @@
+#### THESE ARE TEMPORARY STUBBED APIS ----- REMOVE WHEN THE REAL ONES BECOME AVAILABLE
