+++ stubbed-apis/dataset-api-stubber/dataset.json
@@ -1,0 +1,14 @@
+{
+  "id": "3784782",
+  "title": "Consumer Price Index",
+  "url": "/datasets/3784782/editions/2017/versions/1",
+  "release_date": "11 Nov 2017",
+  "next_release": "11 Nov 2019",
+  "edition": "2017",
+  "version": "1",
+  "contact": {
+    "name": "Matt Rout",
+    "telephone": "07984598308",
+    "email": "matt@gmail.com"
+  }
+}
