+++ stubbed-apis/filter-api-stubber/goods-and-services-options.json
@@ -1,0 +1,17 @@
+{
+  "items": [
+    {
+      "id": "3eygduetydueh",
+      "name": "Clothing"
+    },
+    {
+      "id": "3eysduetydueh",
+      "name": "Confectionary"
+    },
+    {
+      "id": "3eygduetydufh",
+      "name": "Groceries"
+    }
+  ],
+  "number_of_results": 3
+}
