+++ stubbed-apis/filter-api-stubber/job.json
@@ -1,0 +1,6 @@
+{
+  "filter_id": "1235281",
+  "dataset": "2894982",
+  "edition": "2017",
+  "version": "1"
+}
