+++ stubbed-apis/filter-api-stubber/time-options.json
@@ -1,0 +1,17 @@
+{
+  "items": [
+    {
+      "id": "3eygduetydueh",
+      "name": "1996.01"
+    },
+    {
+      "id": "3eysduetydueh",
+      "name": "1996.05"
+    },
+    {
+      "id": "3eygduetydufh",
+      "name": "2000.03"
+    }
+  ],
+  "number_of_results": 3
+}
