+++ stubbed-apis/filter-api-stubber/time.json
@@ -1,0 +1,4 @@
+{
+  "name": "time",
+  "dimension_url": "/code-lists/64d384f1-ea3b-445c-8fb8-aa453f96e58a/codes"
+}
